#24 Add ability to edit your own posts
