+++ app/serializers.py
@@ -25,4 +25,8 @@ class AuthorSerializers(serializers.ModelSerializer):
 class PostSerializer(serializers.ModelSerializer):
     class Meta:
         model = Post
-        fields = ('id', 'author', 'privacy', 'text', 'date_created')
+        fields = ('id', 'author', 'privacy', 'title', 'text', 'date_created')
+
+    def update(self, instance, validated_data):
+        print(validated_data)
+        x = validated_data
