+++ app/templates/posts/create_post.html
@@ -6,7 +6,7 @@
 {% endblock %}
 
 {% block content %}
-    <h1 class="h3 mb-4 text-gray-800">Edit Post</h1>
+    <h1 class="h3 mb-4 text-gray-800">Create Post</h1>
     <div>
         <form action="" method="POST">
             {% csrf_token %}
