+++ app/templates/posts/post_delete.html
@@ -11,7 +11,7 @@
         {% csrf_token %}
         Are you sure you want to delete this post?
         <input class="btn btn-primary" type="submit" value="Yes"/>
-        <a href='{% url 'app:myposts' %}'>Cancel</a>
+        <a href='{% url 'app:my_posts' %}'>Cancel</a>
     </form>
 
 {% endblock %}
