+++ app/urls.py
@@ -1,6 +1,8 @@
 from django.urls import path, include
 from rest_framework import routers
-from app.views import api_views, views
+
+import app.views.post_views
+from app.views import api_views, views, post_views
 from django.conf.urls import url
 
 app_name = 'app'
@@ -19,9 +21,10 @@ urlpatterns = [
     path('register', views.register_view, name="register"),
     path('uploadimage', views.upload_image_view, name="uploadimage"),
     path('images/<str:username>/<str:filename>/<str:encoding>', views.get_image, name="images"),
-    path('my-posts', views.my_posts_view, name="my_posts"),
-    path('create-post', views.create_post_view, name='create_post'),
-    url(r'^delete/(?P<id>\d+)/$', views.delete_post, name="delete"),
+    path('my-posts', app.views.post_views.my_posts_view, name="my_posts"),
+    path('create-post', app.views.post_views.create_post_view, name='create_post'),
+    url(r'^delete/(?P<id>\d+)/$', app.views.post_views.delete_post, name="delete"),
     path('profile', views.profile_view, name="profile"),
     path('profile/edit/', views.edit_profile, name="edit_profile"),
+    path('post/edit/<int:id>/', post_views.edit_post, name="edit_post")
 ]
